+++ src/main/java/shop/readmecorp/adminserverreadme/modules/publisher/dto/PublisherDTO.java
@@ -15,8 +15,6 @@ public class PublisherDTO {
 
     private String username;
 
-    private String password;
-
     private String role;
 
     private String businessNumber;
