+++ src/main/java/shop/readmecorp/adminserverreadme/modules/publisher/entity/Publisher.java
@@ -62,10 +62,10 @@ public class Publisher extends BaseTime {
     }
 
     public PublisherDTO toDTO() {
-        return new PublisherDTO(id, username, password, role.name(),businessNumber, businessName, joinTime.toString(), status.name() );
+        return new PublisherDTO(id, username,role.name(),businessNumber, businessName, joinTime.toString(), status.name() );
     }
 
     public PublisherResponse toResponse() {
-        return new PublisherResponse(id, username, password, role.name(),businessNumber, businessName, joinTime.toString(), status.name());
+        return new PublisherResponse(id, username,role.name(),businessNumber, businessName, joinTime.toString(), status.name());
     }
 }
