+++ src/main/java/shop/readmecorp/adminserverreadme/modules/publisher/response/PublisherResponse.java
@@ -15,8 +15,6 @@ public class PublisherResponse {
 
     private String username;
 
-    private String password;
-
     private String role;
 
     private String businessNumber;
