+++ src/main/webapp/WEB-INF/view/publisher/bookmanage/bookList.jsp
@@ -19,58 +19,58 @@
             </thead>
             <tbody>
 <%-- --%>
-<tr>
-    <th scope="row">1</th>
-    <td><img src="/images/gray.png" style="width: 75px;height: 100px"></td>
-    <td><a href="/books/detail/1">1984</a></td>
-    <td>조지오웰</td>
-    <td>4.5</td>
-    <td>100</td>
-    <td> - </td>
-    <td>
-        <a href="/books/updateForm/1"><button type="button" class="btn btn-outline-dark">수정</button></a>
-        <a href="/books/deleteForm/1"><button type="button" class="btn btn-outline-dark">삭제</button></a>
-    </td>
-</tr>
-<tr>
-    <th scope="row">2</th>
-    <td><img src="/images/gray.png" style="width: 75px;height: 100px"></td>
-    <td>1984</td>
-    <td>조지오웰</td>
-    <td>4.5</td>
-    <td>100</td>
-    <td> - </td>
-    <td>
-        <button>수정</button>
-        <button>삭제</button>
-    </td>
-</tr>
-<tr>
-    <th scope="row">3</th>
-    <td><img src="/images/gray.png" style="width: 75px;height: 100px"></td>
-    <td>1984</td>
-    <td>조지오웰</td>
-    <td>4.5</td>
-    <td>100</td>
-    <td> - </td>
-    <td>
-        <button>수정</button>
-        <button>삭제</button>
-    </td>
-</tr>
-<tr>
-    <th scope="row">4</th>
-    <td><img src="/images/gray.png" style="width: 75px;height: 100px"></td>
-    <td>1984</td>
-    <td>조지오웰</td>
-    <td>4.5</td>
-    <td>100</td>
-    <td> - </td>
-    <td>
-        <button>수정</button>
-        <button>삭제</button>
-    </td>
-</tr>
+<%--<tr>--%>
+<%--    <th scope="row">1</th>--%>
+<%--    <td><img src="/images/gray.png" style="width: 75px;height: 100px"></td>--%>
+<%--    <td><a href="/books/detail/1">1984</a></td>--%>
+<%--    <td>조지오웰</td>--%>
+<%--    <td>4.5</td>--%>
+<%--    <td>100</td>--%>
+<%--    <td> - </td>--%>
+<%--    <td>--%>
+<%--        <a href="/books/updateForm/1"><button type="button" class="btn btn-outline-dark">수정</button></a>--%>
+<%--        <a href="/books/deleteForm/1"><button type="button" class="btn btn-outline-dark">삭제</button></a>--%>
+<%--    </td>--%>
+<%--</tr>--%>
+<%--<tr>--%>
+<%--    <th scope="row">2</th>--%>
+<%--    <td><img src="/images/gray.png" style="width: 75px;height: 100px"></td>--%>
+<%--    <td>1984</td>--%>
+<%--    <td>조지오웰</td>--%>
+<%--    <td>4.5</td>--%>
+<%--    <td>100</td>--%>
+<%--    <td> - </td>--%>
+<%--    <td>--%>
+<%--        <button>수정</button>--%>
+<%--        <button>삭제</button>--%>
+<%--    </td>--%>
+<%--</tr>--%>
+<%--<tr>--%>
+<%--    <th scope="row">3</th>--%>
+<%--    <td><img src="/images/gray.png" style="width: 75px;height: 100px"></td>--%>
+<%--    <td>1984</td>--%>
+<%--    <td>조지오웰</td>--%>
+<%--    <td>4.5</td>--%>
+<%--    <td>100</td>--%>
+<%--    <td> - </td>--%>
+<%--    <td>--%>
+<%--        <button>수정</button>--%>
+<%--        <button>삭제</button>--%>
+<%--    </td>--%>
+<%--</tr>--%>
+<%--<tr>--%>
+<%--    <th scope="row">4</th>--%>
+<%--    <td><img src="/images/gray.png" style="width: 75px;height: 100px"></td>--%>
+<%--    <td>1984</td>--%>
+<%--    <td>조지오웰</td>--%>
+<%--    <td>4.5</td>--%>
+<%--    <td>100</td>--%>
+<%--    <td> - </td>--%>
+<%--    <td>--%>
+<%--        <button>수정</button>--%>
+<%--        <button>삭제</button>--%>
+<%--    </td>--%>
+<%--</tr>--%>
 <%-- --%>
             </tbody>
         </table>
@@ -80,20 +80,20 @@
 </div>
 
 <script>
-    // $(document).ready(function() {
-    //     $.ajax({
-    //         url: 'http://localhost:8080/api/books',
-    //         type: 'GET',
-    //         dataType: 'json',
-    //
-    //     })
-    //         .done((res) => {
-    //             populateTable(res.content); // 'content' 속성을 사용하도록 수정
-    //         })
-    //         .fail((err) => {
-    //             alert(err.responseJSON.msg);
-    //         })
-    // });
+    $(document).ready(function() {
+        $.ajax({
+            url: 'http://localhost:8080/books',
+            type: 'GET',
+            dataType: 'json',
+
+        })
+            .done((res) => {
+                populateTable(res.content); // 'content' 속성을 사용하도록 수정
+            })
+            .fail((err) => {
+                alert(err.responseJSON.msg);
+            })
+    });
 
     function populateTable(books) {
         var tbody = $('table tbody');
@@ -103,13 +103,13 @@
             var tr = $('<tr>');
             tr.append('<th scope="row">' + (i + 1) + '</th>');
             tr.append('<td><img src="/images/gray.png" style="width: 75px;height: 100px"></td>');
-            tr.append('<td><a href="/books/detail/' + book.id + '">' + book.title + '</a></td>');
+            tr.append('<td><a href="/publishers/books/detail/' + book.id + '">' + book.title + '</a></td>');
             tr.append('<td>' + book.author + '</td>');
             tr.append('<td>' + book.status + '</td>');
             tr.append(`
             <td>
-                <button>수정</button>
-                <button>삭제</button>
+                <a href="/publishers/books/updateForm/`+ book.id + `"><button>수정</button></a>
+                <a href="/publishers/books/deleteForm/`+ book.id + `"><button>삭제</button></a>
             </td>
         `);
             tbody.append(tr);
